exposed globals for extending expo.modules
