+++ packages/expo-haptics/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### 💡 Others
 
+- Use JSI global directly for binding.
 - drop unused web `name` property. ([#27437](https://github.com/expo/expo/pull/27437) by [@EvanBacon](https://github.com/EvanBacon))
 - Removed deprecated backward compatible Gradle settings. ([#28083](https://github.com/expo/expo/pull/28083) by [@kudo](https://github.com/kudo))
