--- packages/expo-haptics/src/ExpoHaptics.ts
@@ -1,3 +1,0 @@
-import { requireNativeModule } from 'expo-modules-core';
-
-export default requireNativeModule('ExpoHaptics');
