--- packages/expo-haptics/src/ExpoHaptics.web.ts
@@ -1,1 +1,0 @@
-export default {};
