+++ packages/expo-modules-core/src/ts-declarations/global.ts
@@ -2,11 +2,128 @@ import type { EventEmitter } from './EventEmitter';
 import type { NativeModule } from './NativeModule';
 import type { SharedObject } from './SharedObject';
 
+declare global {
+  /**
+   * Type definitions for the native modules that are installed in Apple runtime environments.
+   *
+   * ---
+   * **TypeScript types**: The TypeScript types for the Apple native module.
+   * ```ts
+   * declare global {
+   *   namespace AppleRuntime {
+   *     interface NativeAppIcon {
+   *       // Constants
+   *       isSupported: boolean;
+   *
+   *       // Methods
+   *       setAlternateIcon: (iconName: string | null) => Promise<string | null>;
+   *     }
+   *   }
+   * }
+   * ```
+   * ---
+   * **Swift module**: The module definition for the native module.
+   * ```swift
+   * import ExpoModulesCore
+   *
+   * public class ExpoAppIconModule: Module {
+   *     public func definition() -> ModuleDefinition {
+   *         // Sets the name of the module in the global `expo.modules` object
+   *         Name("ExpoAppIcon")
+   *
+   *        // Constants
+   *         Constants([
+   *             "isSupported": UIApplication.shared.supportsAlternateIcons
+   *         ])
+   *
+   *         // Methods
+   *         AsyncFunction("setAlternateIcon") { (name: String?, promise: Promise) in
+   *             // ...
+   *         }
+   *     }
+   * }
+   * ```
+   */
+  namespace AppleRuntime {}
+
+  /**
+   * Type definitions for the native modules that are installed in Android runtime environments.
+   *
+   * ---
+   * **TypeScript types**: The TypeScript types for the Android native module.
+   * ```ts
+   * declare global {
+   *   namespace AndroidRuntime {
+   *     interface NativeAppIcon {
+   *       // Constants
+   *       isSupported: boolean;
+   *
+   *       // Methods
+   *       setAlternateIcon: (iconName: string | null) => Promise<string | null>;
+   *     }
+   *   }
+   * }
+   * ```
+   * ---
+   * **Kotlin module**: The module definition for the native module.
+   * ```kotlin
+   * package expo.modules.haptics
+   *
+   * class ExpoAppIconModule : Module() {
+   *
+   *   override fun definition() = ModuleDefinition {
+   *     Name("ExpoAppIcon")
+   *
+   *     Constants("isSupported" to true)
+   *
+   *     AsyncFunction("setAlternateIcon") { name: String ->
+   *       // ...
+   *     }
+   *   }
+   * }
+   * ```
+   */
+  namespace AndroidRuntime {}
+
+  /**
+   * Interface for defining cross-platform native modules that are available on the global `expo.modules` object in custom native runtimes (iOS, Android clients).
+   * Third-party packages can extend this interface to add their own native modules to the global `expo.modules` object.
+   *
+   * Create a native module with:
+   * **`npx create-expo-module@latest --local`**
+   *
+   * Learn more: [creating a native module](https://docs.expo.dev/modules/get-started/).
+   *
+   *
+   * @example
+   * ```ts
+   * declare global {
+   *   namespace AppleRuntime {
+   *     interface NativeAppIcon {
+   *       setAlternateIcon: (iconName: string | null) => Promise<string | null>;
+   *     }
+   *   }
+   *
+   *   namespace AndroidRuntime {
+   *     interface NativeAppIcon {
+   *       setAlternateIcon: (iconName: string | null) => Promise<string | null>;
+   *     }
+   *   }
+   *
+   *   interface ExpoNativeModules {
+   *     NativeAppIcon: AndroidRuntime.NativeAppIcon | AppleRuntime.NativeAppIcon;
+   *   }
+   * }
+   * ```
+   */
+  interface ExpoNativeModules {}
+}
+
 export interface ExpoGlobal {
   /**
    * Host object that is used to access native Expo modules.
    */
-  modules: Record<string, any>;
+  modules: ExpoNativeModules;
 
   // Natively defined JS classes
